consumer/FilterConsumer/requirements.txt
python-dotenv
kafka-python-ng
pyspark